feat: Initialize monorepo structure with pnpm workspaces
