+++ .vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "bd1e60a5",
+  "configHash": "c1fc73ba",
+  "lockfileHash": "bdaa8e02",
+  "browserHash": "60065936",
+  "optimized": {},
+  "chunks": {}
+}
