+++ packages/api/.env
@@ -1,0 +1,2 @@
+GEMINI_API_KEY=PLACEHOLDER_API_KEY
+PORT=3001
