+++ packages/api/package.json
@@ -1,0 +1,15 @@
+{
+    "name": "@devasign/api",
+    "version": "0.0.0",
+    "scripts": {
+        "build": "tsc",
+        "dev": "ts-node-dev --respawn --transpile-only src/index.ts",
+        "start": "node dist/index.js",
+        "lint": "eslint .",
+        "typecheck": "tsc --noEmit"
+    },
+    "devDependencies": {
+        "ts-node-dev": "^2.0.0",
+        "typescript": "^5.7.2"
+    }
+}
