+++ packages/mobile/public/devasign-black.png
