+++ packages/mobile/public/devasign-white.png
